move server_address.go from pb to util
